services/credit-service/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.29
psycopg[binary]==3.1.18
python-jose==3.3.0
celery[redis]>=5.3.6
setuptools>=59.6.0
click==8.1.8
kombu==5.5.3
python-dotenv
pandas>=2.2
boto3>=1.26.0
pydantic-settings>=2.9.1,<3.0
asyncpg>=0.27.0,<1.0
python-multipart>=0.0.5